+++ src/components/content-types/events-discovery/components/Hit/Hit.jsx
@@ -1,0 +1,4 @@
+import React from 'react';
+import SynchronizedEvent from '../../../synchronized-event/synchronizedEvent';
+
+export const Hit = ({ hit }) => <SynchronizedEvent blok={hit} />;
