+++ src/components/content-types/events-discovery/components/constants.js
@@ -1,0 +1,11 @@
+export const facetLabels = {
+  format: {
+    label: 'Format',
+  },
+  experience: {
+    label: 'Experience',
+  },
+  subject: {
+    label: 'Subject',
+  },
+};
